(1) Factor out all sleep regularity index related calculations into a separate Utils file (2) Make a separate test case for sleep regularity index (3) Revert SocialJetlagStatsTest to original test file
